Added completions for commands.
Moved structopt Opt to cli.rs
Add build.rs
