+++ Cargo.toml
@@ -8,6 +8,17 @@ keywords = ["system", "fetch", "cli"]
 repository = "https://github.com/Macchina-CLI/macchina"
 license = "MIT"
 readme = "README_CARGO.md"
+build = "build.rs"
+
+[build-dependencies]
+libmacchina = "0.2.3"
+bytesize = "1.0.1"
+clap = "2.33.3"
+tui = { version = "0.14.0", default-features = false }
+colored = "2.0.0"
+structopt = "0.3.21"
+rand = "0.8.3"
+unicode-width = "0.1.8"
 
 [profile.release]
 opt-level = 3
@@ -26,9 +37,11 @@ structopt = "0.3.21"
 rand = "0.8.3"
 unicode-width = "0.1.8"
 lazy_static = "1.4.0"
+macchina = { path = "." }
 
 [target.'cfg(not(target_env = "msvc"))'.dependencies]
 jemallocator = "0.3.2"
+
 
 [features]
 openwrt = ["libmacchina/openwrt"]
