+++ src/main.rs
@@ -1,9 +1,9 @@
 mod bars;
+mod cli;
 mod format;
 mod theme;
 
-use clap::arg_enum;
-use clap::crate_authors;
+use cli::{MacchinaColor, Opt};
 use std::io;
 use structopt::StructOpt;
 #[macro_use]
@@ -33,184 +33,6 @@ use tui::style::Color;
 use tui::text::Text;
 use tui::widgets::{Block, BorderType, Borders, Paragraph, Widget};
 use unicode_width::UnicodeWidthStr;
-
-pub const AUTHORS: &str = crate_authors!();
-pub const ABOUT: &str =
-    "A system information fetcher, with an emphasis on performance and minimalism.";
-
-arg_enum! {
-    #[derive(Debug)]
-    pub enum MacchinaColor {
-        Red,
-        Green,
-        Blue,
-        Yellow,
-        Cyan,
-        Magenta,
-        Black,
-        White
-    }
-}
-
-impl MacchinaColor {
-    /// Convert the argument passed to a color flag to its respective color.
-    fn get_color(&self) -> Color {
-        match self {
-            MacchinaColor::Red => Color::Red,
-            MacchinaColor::Green => Color::Green,
-            MacchinaColor::Blue => Color::Blue,
-            MacchinaColor::Yellow => Color::Yellow,
-            MacchinaColor::Cyan => Color::Cyan,
-            MacchinaColor::Magenta => Color::Magenta,
-            MacchinaColor::Black => Color::Black,
-            MacchinaColor::White => Color::White,
-        }
-    }
-}
-
-#[derive(StructOpt, Debug)]
-#[structopt(author = AUTHORS, about = ABOUT)]
-pub struct Opt {
-    #[structopt(short = "p", long = "palette", help = "Displays color palette")]
-    palette: bool,
-
-    #[structopt(
-        short = "P",
-        long = "padding",
-        help = "Specify the amount of left padding to use"
-    )]
-    padding: Option<usize>,
-
-    #[structopt(
-        short = "s",
-        long = "spacing",
-        help = "Specify the amount of spacing to use"
-    )]
-    spacing: Option<usize>,
-
-    #[structopt(short = "n", long = "no-color", help = "Disables color")]
-    no_color: bool,
-
-    #[structopt(
-        short = "K",
-        long = "no-separator",
-        help = "Hides the separator",
-        conflicts_with = "separator_color"
-    )]
-    no_separator: bool,
-
-    #[structopt(
-        short = "D",
-        long = "no-bar-delimiter",
-        help = "Hides the bar's delimiters"
-    )]
-    no_bar_delimiter: bool,
-
-    #[structopt(
-    short = "c",
-    long = "color",
-    possible_values = & MacchinaColor::variants(),
-    case_insensitive = true,
-    help = "Specify the key color",
-    conflicts_with = "no_color",
-    )]
-    color: Option<MacchinaColor>,
-
-    #[structopt(
-        short = "b",
-        long = "bar",
-        help = "Displays bars instead of numerical values"
-    )]
-    bar: bool,
-
-    #[structopt(
-    short = "C",
-    long = "separator-color",
-    possible_values = & MacchinaColor::variants(),
-    case_insensitive = true,
-    help = "Specify the separator color",
-    conflicts_with = "no_color",
-    )]
-    separator_color: Option<MacchinaColor>,
-
-    #[structopt(
-        short = "r",
-        long = "random-color",
-        help = "Picks a random key color for you"
-    )]
-    random_color: bool,
-
-    #[structopt(
-        short = "R",
-        long = "random-sep-color",
-        help = "Picks a random separator color for you"
-    )]
-    random_sep_color: bool,
-
-    #[structopt(
-    short = "H",
-    long = "hide",
-    possible_values = & data::ReadoutKey::variants(),
-    case_insensitive = true,
-    help = "Hides the specified elements",
-    min_values = 1,
-    conflicts_with = "show_only"
-    )]
-    hide: Option<Vec<data::ReadoutKey>>,
-
-    #[structopt(
-    short = "X",
-    long = "show-only",
-    possible_values = & data::ReadoutKey::variants(),
-    case_insensitive = true,
-    help = "Displays only the specified elements",
-    min_values = 1,
-    conflicts_with = "hide"
-    )]
-    show_only: Option<Vec<data::ReadoutKey>>,
-
-    #[structopt(short = "d", long = "doctor", help = "Checks the system for failures")]
-    doctor: bool,
-
-    #[structopt(short = "U", long = "short-uptime", help = "Shortens uptime output")]
-    short_uptime: bool,
-
-    #[structopt(short = "S", long = "short-shell", help = "Shortens shell output")]
-    short_shell: bool,
-
-    #[structopt(
-    short = "t",
-    long = "theme",
-    default_value = "Hydrogen",
-    possible_values = & theme::Themes::variants(),
-    case_insensitive = true,
-    help = "Specify the theme"
-    )]
-    theme: theme::Themes,
-
-    #[structopt(long = "no-ascii", help = "Removes the ascii art")]
-    no_ascii: bool,
-
-    #[structopt(
-        long = "no-box",
-        help = "Removes the box surrounding system information"
-    )]
-    no_box: bool,
-
-    #[structopt(
-        long = "box-title",
-        help = "Overrides the title of the box",
-        conflicts_with = "no_box"
-    )]
-    box_title: Option<String>,
-
-    #[structopt(
-        long = "no-title",
-        help = "Hides the box title",
-        conflicts_with = "box_title"
-    )]
-    no_title: bool,
-}
 
 fn create_backend() -> CrosstermBackend<Stdout> {
     CrosstermBackend::new(io::stdout())
@@ -311,7 +133,7 @@ fn create_theme(opt: &Opt) -> Box<dyn Theme> {
 
     if opt.no_bar_delimiter {
         let new_bar = theme.get_bar_style().hide_delimiters();
-        theme.set_bar_style(new_bar.clone());
+        theme.set_bar_style(new_bar);
     }
 
     if opt.random_color {
@@ -407,39 +229,39 @@ fn write_buffer_to_console(
         find_last_buffer_cell_index(tmp_buffer).expect("Error while writing to terminal buffer.");
 
     print!("{}", "\n".repeat(last_y as usize + 1));
-
+    let cursor_y: u16;
     if atty::is(Stream::Stdout) {
-        let (_, cursor_y) = backend.get_cursor().unwrap_or((0, 0));
-        let term_size = backend.size().unwrap_or_default();
-        // We need a checked subtraction here, because (cursor_y - last_y - 1) might underflow if the
-        // cursor_y is smaller than (last_y - 1).
-        let starting_pos = cursor_y.saturating_sub(last_y).saturating_sub(1);
-        let mut skip_n = 0;
-
-        let iter = tmp_buffer
-            .content
-            .iter()
-            .enumerate()
-            .filter(|(_previous, cell)| {
-                let curr_width = cell.symbol.width();
-                if curr_width == 0 {
-                    return false;
-                }
-
-                let old_skip = skip_n;
-                skip_n = curr_width.saturating_sub(1);
-                return old_skip == 0;
-            })
-            .map(|(idx, cell)| {
-                let (x, y) = tmp_buffer.pos_of(idx);
-                (x, y, cell)
-            })
-            .filter(|(x, y, _)| *x < term_size.width && *y <= last_y)
-            .map(|(x, y, cell)| (x, y + starting_pos, cell))
-            .into_iter();
-
-        backend.draw(iter)?;
+        cursor_y = backend.get_cursor().unwrap_or((0, 0)).1;
+    } else {
+        cursor_y = 0
     }
+    let term_size = backend.size().unwrap_or_default();
+    // We need a checked subtraction here, because (cursor_y - last_y - 1) might underflow if the
+    // cursor_y is smaller than (last_y - 1).
+    let starting_pos = cursor_y.saturating_sub(last_y).saturating_sub(1);
+    let mut skip_n = 0;
 
+    let iter = tmp_buffer
+        .content
+        .iter()
+        .enumerate()
+        .filter(|(_previous, cell)| {
+            let curr_width = cell.symbol.width();
+            if curr_width == 0 {
+                return false;
+            }
+
+            let old_skip = skip_n;
+            skip_n = curr_width.saturating_sub(1);
+            old_skip == 0
+        })
+        .map(|(idx, cell)| {
+            let (x, y) = tmp_buffer.pos_of(idx);
+            (x, y, cell)
+        })
+        .filter(|(x, y, _)| *x < term_size.width && *y <= last_y)
+        .map(|(x, y, cell)| (x, y + starting_pos, cell));
+
+    backend.draw(iter)?;
     Ok(())
 }
